+++ all-is-cubes-desktop/src/config_files.rs
@@ -1,15 +1,17 @@
-use std::fs::create_dir_all;
+use std::fs;
 use std::fs::File;
 use std::io::BufReader;
 use std::path::Path;
+use std::sync::Arc;
 
+use all_is_cubes_ui::apps::Settings;
 use directories_next::ProjectDirs;
 use serde::{de::DeserializeOwned, Serialize};
 
 use all_is_cubes_render::camera::GraphicsOptions;
 
 /// Load preferences/settings/config files from a platform-appropriate read/write location.
-pub fn load_config() -> Result<GraphicsOptions, anyhow::Error> {
+pub fn load_config() -> Result<Settings, anyhow::Error> {
     // TODO: make testable
     // TODO: allow users of this library function to pick their own config dir
 
@@ -23,15 +25,23 @@ pub fn load_config() -> Result<GraphicsOptions, anyhow::Error> {
 
     let project_dirs = ProjectDirs::from("org.switchb", "", "all-is-cubes")
         .ok_or_else(|| anyhow::anyhow!("could not find configuration directory"))?;
-    create_dir_all(project_dirs.config_dir())?;
+    fs::create_dir_all(project_dirs.config_dir())?;
+
+    let settings_path = project_dirs.config_dir().join("graphics.json");
 
     let graphics_options = read_or_create_default_json_file(
         "graphics options",
-        &project_dirs.config_dir().join("graphics.json"),
+        &settings_path,
         GraphicsOptions::default,
     );
 
-    Ok(graphics_options)
+    Ok(Settings::with_persistence(
+        Arc::new(graphics_options),
+        // TODO: ideally, writes would be performed asynchronously and with rate-limiting
+        Arc::new(move |data| {
+            write_json_file("graphics options", &settings_path, &data);
+        }),
+    ))
 }
 
 fn read_or_create_default_json_file<V: DeserializeOwned + Serialize>(
@@ -63,7 +73,7 @@ fn read_or_create_default_json_file<V: DeserializeOwned + Serialize>(
             );
             let value = default();
             let json_text = serde_json::to_string_pretty(&value).unwrap();
-            std::fs::write(path, json_text.as_bytes()).expect("Error writing default file");
+            fs::write(path, json_text.as_bytes()).expect("Error writing default file");
             value
         }
         Err(e) => {
@@ -74,6 +84,30 @@ fn read_or_create_default_json_file<V: DeserializeOwned + Serialize>(
                 e
             );
             default()
+        }
+    }
+}
+
+fn write_json_file<V: Serialize>(description: &str, path: &Path, value: &V) {
+    match fs::OpenOptions::new().write(true).truncate(true).open(path) {
+        Ok(file) => match serde_json::to_writer_pretty(file, &value) {
+            Ok(()) => log::trace!("Wrote {} to {}", description, path.to_string_lossy()),
+            Err(e) => {
+                log::error!(
+                    "Error while writing {} file {}: {}",
+                    description,
+                    path.to_string_lossy(),
+                    e
+                );
+            }
+        },
+        Err(e) => {
+            log::error!(
+                "Error while opening {} file {}: {}",
+                description,
+                path.to_string_lossy(),
+                e
+            );
         }
     }
 }