+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Functionality:
+    - In the desktop version, changes to graphics options are now persisted across launches.
+
 - `all-is-cubes-ui` library:
     - `apps::Settings` manages user-editable settings that eventually will be more than just the graphics options.
     - `apps::SessionBuilder::settings()` links a possibly-shared `Settings` to the created session.